Temporary fix issue #3
This temporarily fixes issue #3 but for now user can't be told about
this failure. It silently ignores the error and moves on. When user
tries to use some modules error will be logged in actual silpa.log.

We need to find a way to fix this more cleanly as this will be a dirty
fix