Add method invoking logic to JSON-RPC
Logic for invoking requested method on a module is added, if module
doesn't have a method by the requested name a METHOD_NOT_FOUND error is
returned to client as per JSON-RPC 2.0 specifications.

If there is error in execution of call, INTERNAL_ERROR is returned to
client as per JSON-RPC spec and message is set appropriately to reflect
what happened.

Signed-off-by: Vasudev Kamath <kamathvasudev@gmail.com>